Fixe divers bug for build

+++ Source/GoodManStory/Character/Players/BasePlayer.h
@@ -374,7 +374,7 @@ public:
     void SetRage(float NewRage) noexcept;
 
     UFUNCTION(BlueprintCallable, Category = Stats)
-    void SetLevel(float NewLevel) noexcept;
+    void SetLevel(int NewLevel) noexcept;
     
     UFUNCTION(BlueprintCallable, Category = Stats)
     void TakeRage(float AdditionnalRage) noexcept;
